[Android] Remove the way to enable remote debugging via Intent
The remote debugging could be enabled via intent for apps created
by make_apk, which is not what developers want.

Remove this feature, now only runtime client shells can turn on
remote debugging in this way.

For app developers, they can use make_apk's arguments to enable
remote debugging if they want.

BUG=https://crosswalk-project.org/jira/browse/XWALK-2778
